movies_admin/requirements.txt
django==3.2
python_dotenv==1.0.0
psycopg2==2.9.7
django-split-settings==1.2.0
python_dateutil==2.8.2
djangorestframework==3.14.0
uwsgi==2.0.22